feat(frontend): Diff modal
